+++ datastock/_class01_interpolate.py
